Fix delete category crash
Application would crash when user deleted category because the ID of the category in the database was missing. When not all the IDs are consecutive, the for loop would try reading an item that wasn't there. In this commit, I reset the IDs of the remaining categories after deleting the selected one.